requirements.txt
coloredlogs==0.8
websocket-client==0.31.0
